refactor: split detector implementations into dedicated modules with barrel exports
